Added pagination support for group-management for major db engines.
